Edured-84: Added flow for request deactivation when trying to start an enrollment after web sign in.
Before calling startEnrollment endpoint there is now a check to see if the current account for which authentication was  just completed includes an app as a login option. If the login options already include an app, that means a TIQR enrollment was already done for this account. In order to continue enrolling the current app a deactivation of the existing enrollment must be performed first.

Change confirm code screen to allow deactivation on top of phone recovery confirmation.

Merged RecoveryRepository.kt into PersonalInfoRepository for easier reuse.