Bugfix/tiqr 359 tiqr 383 refactor to data integration (#126)
* Added the strings from the core module to the localicious.yaml file for eduid app.
With the migration to AGP 8+ and the use of transient classes, we either have to use the strings from the core namespace or put the required strings into the localicious file. Chose the later, despite this currently leading to duplication. In the future the core module dependency will be removed as it mostly contains the old view UI and eduid will rely only on the data module

* Build only bundle and upload only bundle. Doing both APK and bundle is redundant and wasteful.
AAB files cannot be sideloaded, archiving them is not useful.

* Fix R8 full mode enabled

* Remove core dependency

* Move core dependency to data dependency.
Update to AGP8+.

* Fix proguard rules & camera dependencies

* Update core to point to develop i.s.o. dev branch

* Fix the package import for utility functions

* Fix release:
* Make sure debug network configuration is only used for debug app.
* Ensure backup & restore rules are configured for the nl.eduid app specifically.
* Removed redundant color usage in theme, because the xml theme is not used in the app, outside for the window decoration attributes.
* Update some dependencies
* Fixed account created redirect links: server is using production redirect link, not environment dependent, and it's using a custom scheme instead of `https`. Keeping all 3 versions to avoid future headaches.

---------

Co-authored-by: Iulia STANA <iulia@egeniq.com>